requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.2.0
colorama==0.4.6
flasgger==0.9.7.1
Flask==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
mistune==3.1.3
mysql-connector-python==9.3.0
packaging==25.0
python-dotenv==1.1.0
PyYAML==6.0.2
referencing==0.36.2
rpds-py==0.25.0
six==1.17.0
Werkzeug==3.1.3
